+++ cplusplus/include/rabbitizer.hpp
@@ -9,6 +9,8 @@
 #include "common/RabbitizerVersion.h"
 #include "common/RabbitizerConfig.h"
 
+#include "common/Utils.hpp"
+
 #include "instructions/OperandType.hpp"
 #include "instructions/InstrId.hpp"
 //#include "instructions/InstrSuffix.hpp"