+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `flag_disasmAsData` member to the `Instruction` class.
+  - Flag to override the disasmAsData global configuration.
+  - `TrinaryValue.TRUE` forces the instruction to be disassembled as data.
+  - `RAB_TRINARY_VAL_FALSE` bypasses the global checks for disassembling a word
+    as data. A word will still be disassembled as data if it can't be decoded.
+  - `RAB_TRINARY_VAL_NONE` leaves this decision to the global settings.
+
 ## [1.7.10] - 2023-09-24
 
 ### Added